Avatar Creation Updates
The base image of the avatar is now drawn on the canvas using paths. It
is no longer an jpg. The scale and position have changed slightly.

I figured out how to correctly draw multiple images on the answer
canvas. The problem before was that the scale and transform applied to
the individual elements (like shirt) was being added to the scale and
transform applied to the elements placed on the canvas afterward. The
solution is to include Movie.ctxScratch.save() before the canvas
functions drawing a shape and Movie.ctxScratch.restore() after the
canvas functions.

I had to modify the index.html file to allow testing of the various
applications. I needed the “Clear Data” function.