3-24-2015
- Decomposed the shape challenge into three different challenges, the
first of which has the girl assemble the given blocks to draw a square
(sequentiality). The second asks the girl to use a loop to draw a
square. The thirds asks the girl to draw a hexagon using either
sequential blocks or loops.
- Added a slideshow to the homepage and redesigned it. Working on
improvements to the carousel (buttons and secondary navigation). The
page also now loads super slowly, so I’m working on improving this as
well.
- Changed the header sitewide. Logo is a link back to the homepage.
- Cleaned up code in the movie.js file. Removed unnecessary functions
to simplify the file.