Index Page Modifications
The page is still loading slowly (I think this has to do with the
boot.js file), but I changed the login button from an input element to
a button element so it inherits the CSS styles in common.css and
looks/acts like the other buttons in the application. I also changed
the styling of the slideshow so it looks nicer.