Avatar creation updates
Modified the avatar creation logic so that a user can add accessories
within the category for that level and change the color of the
accessory as long as they want. They then click the button “unlock
accessories” to play a challenge, instead of forcing them to play a
challenge as soon as they place an accessory on the avatar. I also
added links for the avatar creation levels. I need to clean up the
movie.js code and styling.